+++ My project/Assets/Scripts/NeuralNetworks/Activations/ReLUActivation.cs
@@ -8,11 +8,11 @@ namespace Assets.Scripts.NeuralNetworks.Activations
 {
     internal class ReLUActivation : ActivationFunction
     {
-        public override double Activate(double value)
+        public override float Activate(float value)
         {
             return value > 0 ? value : 0;
         }
-        public override double Derivative(double value)
+        public override float Derivative(float value)
         {
             return value > 0 ? 1 : 0;
         }
