+++ My project/Assets/Scripts/NeuralNetworks/WeightInits/HeInitializer.cs
@@ -5,16 +5,16 @@ namespace Assets.Scripts.NeuralNetworks.WeightInits
 {
     public class HeInitializer : WeightInitializer
     {
-        public override double[] GenerateWeights(int inputSize)
+        public override float[] GenerateWeights(int inputSize)
         {
-            double stddev = Math.Sqrt(2.0 / inputSize);
+            float stddev = (float)Math.Sqrt(2.0 / inputSize);
             return Enumerable.Range(0, inputSize)
-                .Select(_ => (double)UnityEngine.Random.Range((float)(-stddev), (float)(stddev)))
+                .Select(_ => UnityEngine.Random.Range(-stddev, stddev))
                 .ToArray();
            
         }
 
-        public override double GenerateBias()
+        public override float GenerateBias()
         {
             return 0;
         }
